Move classes into their own packages
