+++ src/main/java/org/opentripplanner/apis/vectortiles/model/MapboxStyleJson.java
@@ -12,11 +12,7 @@ public final class MapboxStyleJson {
   private final List<TileSource> sources;
   private final List<JsonNode> layers;
 
-  public MapboxStyleJson(
-    String name,
-    List<TileSource> sources,
-    List<JsonNode> layers
-  ) {
+  public MapboxStyleJson(String name, List<TileSource> sources, List<JsonNode> layers) {
     this.name = name;
     this.sources = sources;
     this.layers = layers;
