+++ src/main/java/org/opentripplanner/astar/model/BinHeap.java
@@ -79,14 +79,6 @@ public class BinHeap<T> {
     prio[i] = p;
   }
 
-  public void dump() {
-    for (int i = 0; i <= capacity; i++) {
-      String topMarker = (i > size) ? "(UNUSED)" : "";
-      System.out.printf("%d\t%f\t%s\t%s\n", i, prio[i], elem[i], topMarker);
-    }
-    System.out.printf("-----------------------\n");
-  }
-
   public void reset() {
     // empties the queue in one operation
     size = 0;
@@ -134,9 +126,5 @@ public class BinHeap<T> {
     this.capacity = capacity;
     prio = Arrays.copyOf(prio, capacity + 1);
     elem = Arrays.copyOf(elem, capacity + 1);
-  }
-
-  public int getCapacity() {
-    return capacity;
   }
 }
