+++ src/main/java/org/opentripplanner/inspector/vector/edge/EdgeLayerBuilder.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.inspector.vector.geofencing;
+package org.opentripplanner.inspector.vector.edge;
 
 import java.util.List;
 import org.locationtech.jts.geom.Envelope;
@@ -8,35 +8,31 @@ import org.opentripplanner.inspector.vector.LayerBuilder;
 import org.opentripplanner.inspector.vector.LayerParameters;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.index.StreetIndex;
-import org.opentripplanner.service.vehiclerental.street.NoRestriction;
+import org.opentripplanner.street.model.edge.Edge;
 import org.opentripplanner.street.model.vertex.Vertex;
 import org.opentripplanner.transit.model.site.AreaStop;
 
 /**
  * A vector tile layer containing all {@link AreaStop}s inside the vector tile bounds.
  */
-public class GeofencingZonesLayerBuilder extends LayerBuilder<Vertex> {
+public class EdgeLayerBuilder extends LayerBuilder<Edge> {
 
   private final StreetIndex streetIndex;
 
-  public GeofencingZonesLayerBuilder(Graph graph, LayerParameters layerParameters) {
-    super(
-      new GeofencingZonesPropertyMapper(),
-      layerParameters.name(),
-      layerParameters.expansionFactor()
-    );
+  public EdgeLayerBuilder(Graph graph, LayerParameters layerParameters) {
+    super(new EdgePropertyMapper(), layerParameters.name(), layerParameters.expansionFactor());
     this.streetIndex = graph.getStreetIndex();
   }
 
   @Override
   protected List<Geometry> getGeometries(Envelope query) {
     return streetIndex
-      .getVerticesForEnvelope(query)
+      .getEdgesForEnvelope(query)
       .stream()
-      .filter(se -> !(se.rentalRestrictions() instanceof NoRestriction))
-      .map(vertex -> {
-        Geometry geometry = GeometryUtils.getGeometryFactory().createPoint(vertex.getCoordinate());
-        geometry.setUserData(vertex);
+      .filter(edge -> edge.getGeometry() != null)
+      .map(edge -> {
+        Geometry geometry = edge.getGeometry().copy();
+        geometry.setUserData(edge);
         return geometry;
       })
       .toList();
