+++ src/main/java/org/opentripplanner/inspector/vector/stop/AreaStopsLayerBuilder.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.inspector.vector;
+package org.opentripplanner.inspector.vector.stop;
 
 import java.util.Collection;
 import java.util.List;
@@ -6,7 +6,8 @@ import java.util.Locale;
 import java.util.function.Function;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
-import org.opentripplanner.api.mapping.PropertyMapper;
+import org.opentripplanner.inspector.vector.LayerBuilder;
+import org.opentripplanner.inspector.vector.LayerParameters;
 import org.opentripplanner.transit.model.site.AreaStop;
 import org.opentripplanner.transit.model.site.StopLocation;
 import org.opentripplanner.transit.service.TransitService;
@@ -24,7 +25,7 @@ public class AreaStopsLayerBuilder extends LayerBuilder<StopLocation> {
     Locale locale
   ) {
     super(
-      new DebugClientAreaStopPropertyMapper(locale),
+      new StopLocationPropertyMapper(locale),
       layerParameters.name(),
       layerParameters.expansionFactor()
     );
@@ -38,20 +39,9 @@ public class AreaStopsLayerBuilder extends LayerBuilder<StopLocation> {
       .stream()
       .map(areaStop -> {
         Geometry geometry = areaStop.getGeometry().copy();
-
         geometry.setUserData(areaStop);
-
         return geometry;
       })
       .toList();
-  }
-
-  enum MapperType {
-    DebugClient,
-  }
-
-  @FunctionalInterface
-  private interface MapperFactory {
-    PropertyMapper<AreaStop> build(TransitService transitService, Locale locale);
   }
 }
