+++ src/main/java/org/opentripplanner/inspector/vector/stop/StopLocationPropertyMapper.java
@@ -1,26 +1,27 @@
-package org.opentripplanner.inspector.vector;
+package org.opentripplanner.inspector.vector.stop;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
 import org.opentripplanner.api.mapping.I18NStringMapper;
 import org.opentripplanner.api.mapping.PropertyMapper;
+import org.opentripplanner.inspector.vector.KeyValue;
 import org.opentripplanner.transit.model.site.StopLocation;
 import org.opentripplanner.transit.service.TransitService;
 
 /**
  * A {@link PropertyMapper} for the {@link AreaStopsLayerBuilder} for the OTP debug client.
  */
-public class DebugClientAreaStopPropertyMapper extends PropertyMapper<StopLocation> {
+public class StopLocationPropertyMapper extends PropertyMapper<StopLocation> {
 
   private final I18NStringMapper i18NStringMapper;
 
-  public DebugClientAreaStopPropertyMapper(Locale locale) {
+  public StopLocationPropertyMapper(Locale locale) {
     this.i18NStringMapper = new I18NStringMapper(locale);
   }
 
   public static PropertyMapper<StopLocation> create(TransitService ignored, Locale locale) {
-    return new DebugClientAreaStopPropertyMapper(locale);
+    return new StopLocationPropertyMapper(locale);
   }
 
   @Override
